refactor(sampling): isolate result containers
